feat(main): #56 remove simple-git (#64)
Replace simple-git with custom status parser. This should reduce the size of the project by about 38%. This also fixes a bug where no files were found as staged, or when files differed in the index and work tree. Note, the issue mentions this bug was caused by simple-git. That's not true. It was my own code (surprise surprise). However, the project doesn't need all of what simple-git is.

Closes: #56, #60